接口调试

+++ mall-common/src/main/java/com/shpun/mall/common/service/impl/RedisServiceImpl.java
@@ -1,12 +1,14 @@
 package com.shpun.mall.common.service.impl;
 
 import com.shpun.mall.common.common.Const;
+import com.shpun.mall.common.config.ProfileConfig;
 import com.shpun.mall.common.service.RedisService;
 import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
+import sun.java2d.cmm.Profile;
 
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
@@ -19,8 +21,11 @@ import java.util.concurrent.TimeUnit;
 @Service("redisService")
 public class RedisServiceImpl implements RedisService {
 
-    @Autowired
+    @Autowired(required = false)
     private RedisTemplate<String, Object> redisTemplate;
+
+    @Autowired
+    private ProfileConfig profileConfig;
 
     @Override
     public void set(String key, Object value) {
@@ -44,35 +49,41 @@ public class RedisServiceImpl implements RedisService {
 
     @Override
     public void delete(String key) {
-        redisTemplate.delete(key);
+        if (Const.PROFILE_PROD.equals(profileConfig.getActiveProfile())) {
+            redisTemplate.delete(key);
+        }
     }
 
     @Override
     public void deleteByPrefix(String prefix) {
-        Set<String> keys = redisTemplate.keys(prefix);
-        if (CollectionUtils.isNotEmpty(keys)) {
-            redisTemplate.delete(keys);
+        if (Const.PROFILE_PROD.equals(profileConfig.getActiveProfile())) {
+            Set<String> keys = redisTemplate.keys(prefix);
+            if (CollectionUtils.isNotEmpty(keys)) {
+                redisTemplate.delete(keys);
+            }
         }
     }
 
     @Override
     public void deleteByPrefix(Class cls, String methodName, Object... params) {
-        StringBuilder prefixSb = new StringBuilder(Const.REDIS_KEY_PREFIX)
-                .append(Const.REDIS_KEY_DELIMITER)
-                .append(cls.toString().split("@")[0]);
+        if (Const.PROFILE_PROD.equals(profileConfig.getActiveProfile())) {
+            StringBuilder prefixSb = new StringBuilder(Const.REDIS_KEY_PREFIX)
+                    .append(Const.REDIS_KEY_DELIMITER)
+                    .append(cls.toString().split("@")[0]);
 
-        if (methodName != null) {
-            prefixSb.append(Const.REDIS_KEY_DELIMITER)
-                    .append(methodName);
-        }
-
-        if (params != null) {
-            for (int i = 0; i< params.length; i++) {
-                prefixSb.append(Const.REDIS_PARAM_DELIMITER)
-                        .append(params[i]);
+            if (methodName != null) {
+                prefixSb.append(Const.REDIS_KEY_DELIMITER)
+                        .append(methodName);
             }
-        }
 
-        this.deleteByPrefix(prefixSb.toString());
+            if (params != null) {
+                for (int i = 0; i< params.length; i++) {
+                    prefixSb.append(Const.REDIS_PARAM_DELIMITER)
+                            .append(params[i]);
+                }
+            }
+
+            this.deleteByPrefix(prefixSb.toString());
+        }
     }
 }
